methods refactored as functions
